[ZIL Kinds] move the bulk of LWB-specific code into zil_lwb.c
We are careful to not change anything, particularly not
- the function signatures
- the function bodies
- the relative order of the LWB-specific functions

This makes the diff easy to review:
1. Open the delete-heavy diff for zil.c, and the newly added zil_lwb.c
   side-by-side.
2. Scroll to the top for both the zil.c diff and zil_lwb.c, then go
   through the changes to zil.c top to bottom, performing the following
   check for each function F deleted from zil.c:
   ```
   Search for function's code in zil_lwb.c.
   It should
   - be present verbatim (not even whitespace changes)
   - in the same order relative to the other functions deleted from zil.c
   ```
3. Ensure that no extraneous code was added to zil_lwb.c, i.e., we
   reached the end of zil_lwb.c in step (2).
