+++ module/Makefile.bsd
@@ -363,6 +363,7 @@ CFLAGS.zcp.c= -Wno-cast-qual
 CFLAGS.zfs_fm.c= -Wno-cast-qual
 CFLAGS.zfs_ioctl.c= -Wno-cast-qual
 CFLAGS.zil.c= -Wno-cast-qual
+CFLAGS.zil_lwb.c= -Wno-cast-qual
 CFLAGS.zio.c= -Wno-cast-qual
 CFLAGS.zrlock.c= -Wno-cast-qual
 CFLAGS.zfs_zstd.c= -Wno-cast-qual -Wno-pointer-arith
